check whether output file exists before removal

+++ funtofem/driver/funtofem_shape_driver.py
@@ -287,11 +287,18 @@ class FuntofemShapeDriver(FUNtoFEMnlbgs):
         Create new aero/struct geometries and run fully-coupled forward analysis.
         """
         if self.aero_shape:
+            start_time_aero = time.time()
+            if self.comm.rank == 0:
+                print("F2F - building aero mesh..", flush=True)
             if self.flow_aim.mesh_morph:
                 self.flow_aim.set_design_sensitivity(False, include_file=False)
 
             # run the pre analysis to generate a new mesh
             self.flow_aim.pre_analysis()
+
+            dt_aero = time.time() - start_time_aero
+            if self.comm.rank == 0:
+                print(f"F2F - built aero mesh in {dt_aero:.5e} sec", flush=True)
 
             # for FUN3D mesh morphing now initialize body nodes
             if not (self.is_paired) and self._first_forward:
@@ -307,10 +314,17 @@ class FuntofemShapeDriver(FUNtoFEMnlbgs):
 
         if self.struct_shape:
             # self._update_struct_design()
+            start_time_struct = time.time()
+            if self.comm.rank == 0:
+                print("F2F - Building struct mesh")
             input_dict = {var.name: var.value for var in self.model.get_variables()}
             self.model.structural.update_design(input_dict)
             self.struct_aim.setup_aim()
             self.struct_aim.pre_analysis()
+
+            dt_struct = time.time() - start_time_struct()
+            if self.comm.rank == 0:
+                print(f"F2F - Built struct mesh in {dt_struct:.5e} sec", flush=True)
 
             # move the bdf and dat file to the fun3d_dir
             if self.is_remote:
@@ -341,7 +355,7 @@ class FuntofemShapeDriver(FUNtoFEMnlbgs):
             )
 
             # clear the output file
-            if self.root_proc:
+            if self.root_proc and os.path.exists(self.remote.output_file):
                 os.remove(self.remote.output_file)
 
             start_time = time.time()
